Remove SimpleSplitStringExpression
