+++ stdlib/ballerina-http/src/main/java/org/ballerinalang/net/uri/parser/SimpleStringExpression.java
@@ -21,6 +21,8 @@ package org.ballerinalang.net.uri.parser;
 
 import org.ballerinalang.net.uri.URITemplateException;
 
+import java.util.Map;
+
 /**
  * SimpleStringExpression represents path segments that have single path param.
  * ex - /{foo}/
@@ -28,15 +30,101 @@ import org.ballerinalang.net.uri.URITemplateException;
  * @param <DataType> Type of data which should be stored in the node.
  * @param <InboundMsgType> Inbound message type for additional checks.
  */
-public class SimpleStringExpression<DataType, InboundMsgType>
-        extends SimpleSplitStringExpression<DataType, InboundMsgType> {
+public class SimpleStringExpression<DataType, InboundMsgType> extends Expression<DataType, InboundMsgType> {
+
+    protected static final char[] RESERVED = new char[]{
+            ':', '/', '?', '#', '[', ']', '@', '!', '$', '&', '\'', '(', ')', '*', '+', ',', ';', '='
+    };
 
     public SimpleStringExpression(DataElement<DataType, InboundMsgType> dataElement, String token)
             throws URITemplateException {
         super(dataElement, token);
     }
 
+    @Override
+    String expand(Map<String, String> variables) {
+        boolean emptyString = false;
+        StringBuffer buffer = new StringBuffer();
+        for (Variable var : variableList) {
+            String name = var.getName();
+            if (!variables.containsKey(name)) {
+                continue;
+            }
+            if (buffer.length() > 0) {
+                buffer.append(getSeparator());
+            }
+            String value = var.modify(variables.get(name));
+            if ("".equals(value)) {
+                emptyString = true;
+            }
+            buffer.append(encodeValue(value));
+        }
+
+        if (buffer.length() == 0 && !emptyString) {
+            return null;
+        }
+        return buffer.toString();
+    }
+
+    @Override
+    int match(String uriFragment, Map<String, String> variables) {
+        int length = uriFragment.length();
+        for (int i = 0; i < length; i++) {
+            char ch = uriFragment.charAt(i);
+            if (isEndCharacter(ch)) {
+                if (ch == getSeparator() && variableList.size() > 0) {
+                    continue;
+                }
+
+                if (!setVariables(uriFragment.substring(0, i), variables)) {
+                    return -1;
+                }
+                return i;
+            } else if (i == length - 1) {
+                if (!setVariables(uriFragment, variables)) {
+                    return -1;
+                }
+                return length;
+            }
+        }
+        return 0;
+    }
+
+    @Override
+    char getFirstCharacter() {
+        return '\u0001';
+    }
+
     protected boolean isEndCharacter(Character endCharacter) {
         return endCharacter == '/';
+    }
+
+    protected char getSeparator() {
+        return ',';
+    }
+
+    protected boolean setVariables(String expressionValue, Map<String, String> variables) {
+        String finalValue = decodeValue(expressionValue);
+        for (Variable var : variableList) {
+            String name = var.getName();
+            if (variables.containsKey(name) && !finalValue.equals(variables.get(name))) {
+                return false;
+            }
+            if (var.checkModifier(finalValue)) {
+                variables.put(name, finalValue);
+            } else {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    protected boolean isReserved(char ch) {
+        for (char reservedChar : RESERVED) {
+            if (ch == reservedChar) {
+                return true;
+            }
+        }
+        return false;
     }
 }
