+++ tests/ballerina-test/src/test/resources/testng.xml
@@ -56,7 +56,7 @@ under the License.
             <!--<package name="org.ballerinalang.test.services.basics"/>-->
             <!--<package name="org.ballerinalang.test.services.cors"/>-->
             <!--<package name="org.ballerinalang.test.services.configuration"/>-->
-            <!--<package name="org.ballerinalang.test.services.dispatching"/>-->
+            <package name="org.ballerinalang.test.services.dispatching"/>
             <package name="org.ballerinalang.test.natives.*"/>
             <package name="org.ballerinalang.test.parser.*"/>
             <package name="org.ballerinalang.test.launch.*"/>
